requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2
djangorestframework==3.15.1
dotenv==0.9.9
psycopg2==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.0
typing-extensions==4.13.2
tzdata==2024.1
